services/agent-sofia/requirements.txt
google-cloud-aiplatform[adk,agent_engines]
cloudpickle==3.0
google-genai
pydantic
pytest-asyncio
google-adk
requests
dateparser
google-auth
PyJWT
cryptography
google-generativeai
gunicorn
